% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{first_names_race}
\alias{first_names_race}
\title{Surnames and number of people of each race with that first name}
\format{
A data frame with 4,251 rows and 8 variables:
\describe{
  \item{name}{Surname}
  \item{likely_race}{The most likely race based on the probability of each race}
  \item{probability_american_indian}{Probability that the surname is American Indian}
  \item{probability_asian}{Probability that the surname is Asian}
  \item{probability_black}{Probability that the surname is Black}
  \item{probability_hispanic}{Probability that the surname is Hispanic}
  \item{probability_white}{Probability that the surname is White}
  \item{probability_2races}{Probability that the surname is two or more races}
  ...
}
}
\source{
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/TYJKEZ}
}
\usage{
first_names_race
}
\description{
A dataset containing over 167 thousands surnames and the
number of people of each race with that surname. Citation
for this data: Tzioumis, Konstantinos (2018) Demographic aspects of first names, Scientific Data, 5:180025 [dx.doi.org/10.1038/sdata.2018.25].
}
\keyword{datasets}
