% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{race}
\alias{race}
\title{Find the race of a surname}
\usage{
race(name, probability = TRUE)
}
\arguments{
\item{name}{String or vector of strings of surname that you want to know the race of.}

\item{probability}{If TRUE (default) will provide columns for each race with the probability
that the surname is of that race. If FALSE, will only return the name,
the match-name from the Census data, and the most likely race.}
}
\value{
A data.frame with three columns: The first column has the name as inputted,
the second column has the cleaned up name (no spaces or punctuation,
first letter is capitalized), the final column tells the race of the name
}
\description{
Find the race of a surname
}
\examples{
race("franklin")

race(c("franklin", "Washington", "Jefferson", "Sotomayor", "Liu"))
race("franklin", probability = FALSE)
}
